Move export/import data to bootstrap
